+++ src/fkooman/OAuth/Common/Scope.php
@@ -17,7 +17,7 @@
 
 namespace fkooman\OAuth\Common;
 
-use fkooman\OAuth\Common\Exception\ScopeException;
+use InvalidArgumentException;
 
 class Scope
 {
@@ -28,7 +28,7 @@ class Scope
     {
         foreach ($scope as $s) {
             if (!$this->validateScopeToken($s)) {
-                throw new ScopeException("invalid scope token");
+                throw new InvalidArgumentException("invalid scope token");
             }
         }
         sort($scope, SORT_STRING);
@@ -41,13 +41,13 @@ class Scope
             return new self();
         }
         if (!is_string($scope)) {
-            throw new ScopeException("scope must be string");
+            throw new InvalidArgumentException("scope must be string");
         }
         if (0 === strlen($scope)) {
             return new self();
         }
         if (!is_string($separator)) {
-            throw new ScopeException("separator must be string");
+            throw new InvalidArgumentException("separator must be string");
         }
 
         return new self(explode($separator, $scope));
@@ -143,7 +143,7 @@ class Scope
     public function toString($separator = " ")
     {
         if (!is_string($separator)) {
-            throw new ScopeException("separator must be string");
+            throw new InvalidArgumentException("separator must be string");
         }
 
         return implode($separator, $this->scope);
@@ -165,7 +165,7 @@ class Scope
     private function validateScopeToken($scopeToken)
     {
         if (!is_string($scopeToken) || 0 >= strlen($scopeToken)) {
-            throw new ScopeException("scope token must be a non-empty string");
+            throw new InvalidArgumentException("scope token must be a non-empty string");
         }
         $scopeTokenRegExp = '/^(?:\x21|[\x23-\x5B]|[\x5D-\x7E])+$/';
         $result = preg_match($scopeTokenRegExp, $scopeToken);